Player Sprites Added
Player sprites have been added, together with sprites in general. This is together with an updated rendering system where the system puts the image together, before clearing and then printing the already compiled string. Basically meaning that there is much less stuttering in the rendering process, since it doenst have to process all segments and their functions while drawing to the screen.

New player class aswell, together with sprite class, sprite xml file and parser, the player in itself is not yet added, however it is planned and the infrastructure is in place for its use. Now for future uses it can be used for vars, functions and so on related to plater.

Some small bugs fixed, for example with the screen width