Added draw menu
Added draw menu, which draws a menu and changes the games state.

I also begun working on a  timing system, so that all actions are based on real time and not on the framerate, so that i can finally have a smooth game with a high framerate, with constant character speeds.